repo-size-check 2/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
PyJWT[crypto]==2.8.0
sqlmodel==0.0.14
fastapi-cors==0.0.6
python-dotenv==1.0.0

# Storage providers
boto3
google-api-python-client
google-auth
google-auth-oauthlib
dropbox
requests

# Auth and async
aiohttp
PyYAML

# Logging and utilities
loguru

# Core dependencies only (dev tools in requirements-dev.txt)
